@@ -2,8 +2,8 @@ package com.joniroliveira.bconforto.ui.adapters;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
@@ -12,7 +12,6 @@ import android.widget.TextView;
 import com.joniroliveira.bconforto.data.model.Cloth;
 
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Created by jonioliveira on 18/01/18.
