@@ -10,7 +10,7 @@ import com.joniroliveira.bconforto.data.preferences.Settings;
 
 public class CalculatePrice {
 
-    public static Price calculate(double hours, float hoursPrice, double clothPrice, double quantity, float foamPrice,int discount){
+    public static Price calculate(double hours, float hoursPrice, double clothPrice, double quantity, float foamPrice, int discount){
         double price = (hours * hoursPrice) + (clothPrice * quantity) + foamPrice;
         float aux = (float) (discount/100.0);
         double discountValue = price * aux;
